fix token

+++ src/app/auth/services/token.interceptor.ts
@@ -1,6 +1,6 @@
-import { HttpInterceptorFn } from '@angular/common/http';
+import { HttpInterceptorFn, HttpErrorResponse } from '@angular/common/http';
 import { inject } from '@angular/core';
-import { catchError, throwError } from 'rxjs';
+import { catchError, switchMap, throwError } from 'rxjs';
 import { AuthService } from './auth.service';
 import { Router } from '@angular/router';
 
@@ -9,47 +9,50 @@ export const tokenInterceptor: HttpInterceptorFn = (request, next) => {
   const router = inject(Router);
   const excludedEndpoints: string[] = ['login'];
   const excludedEndpointsAfterRefresh: string[] = ['refresh-token', 'logout'];
-  const tokenData = JSON.parse(localStorage.getItem('tokenData')!);
+  const tokenData = JSON.parse(localStorage.getItem('tokenData') || '{}');
 
   if (excludedEndpoints.some(endpoint => request.url.includes(endpoint))) {
     return next(request);
   }
-  if (tokenData) {
-    const clonedRequest = request.clone({
+
+  if (tokenData && tokenData.token) {
+    request = request.clone({
       setHeaders: {
         Authorization: `Bearer ${tokenData.token}`,
       },
     });
-
-    return next(clonedRequest).pipe(
-      catchError(error => {
-        if (
-          error.status === 401 &&
-          !excludedEndpointsAfterRefresh.some(endpoint =>
-            request.url.includes(endpoint),
-          )
-        ) {
-          authService.refreshToken(tokenData.refreshToken).subscribe({
-            next: (response: any) => {
-              const clonedRequestRepeat = request.clone({
-                url: request.url,
-                setHeaders: {
-                  Authorization: `Bearer ${response.token}`,
-                },
-              });
-              return next(clonedRequestRepeat);
-            },
-            error: () => {
-              localStorage.clear();
-              router.navigate(['auth']);
-              return throwError(() => error);
-            },
-          });
-        }
-        return throwError(() => error);
-      }),
-    );
   }
 
-  return next(request);
+  return next(request).pipe(
+    catchError((error: HttpErrorResponse) => {
+      if (
+        error.status === 401 &&
+        !excludedEndpointsAfterRefresh.some(endpoint =>
+          request.url.includes(endpoint),
+        ) &&
+        tokenData &&
+        tokenData.refreshToken
+      ) {
+        return authService.refreshToken(tokenData.refreshToken).pipe(
+          switchMap((response: any) => {
+            localStorage.setItem('tokenData', JSON.stringify(response));
+            const newRequest = request.clone({
+              setHeaders: {
+                Authorization: `Bearer ${response.token}`,
+              },
+            });
+
+            return next(newRequest);
+          }),
+          catchError(refreshError => {
+            console.error('Error refreshing token:', refreshError);
+            localStorage.clear();
+            router.navigate(['auth']);
+            return throwError(() => refreshError);
+          }),
+        );
+      }
+      return throwError(() => error);
+    }),
+  );
 };
